fixes for new routing logic
